Rework organizers' attribute handling
This commit adds a check which executes when an organizer is called,
which determines if the defined input and success attributes on all
organized interaktors actually "match up".

For now it requires an interaktor's required attributes to be given as
success attributes on the interaktor that immediately preceeds it, but I
think I'll be making a change which will instead check if any of the
previous interaktors defines it as an success attribute.