Validate attributes only on hash input, not in Context
When calling an interaktor you can currently pass in a hash object (the
public interface), or a Context object directly (private API).

Because of this public/private distinction, it makes sense to decouple
the attribute validation logic from the Context object, and instead only
apply it at the point we create a new Context object from a hash object.

Thanks to @blafri on GitHub (https://github.com/blafri) for this
insight, as well as their input and suggestions on changing organizer
logic.